youtube_flatland/utils/flatland_training/requirements.txt
argparse==1.4.0
flatland-rl==2.2.1
numpy==1.18.5
torch==1.5.0
opencv-python==4.2.0.34
Pillow==7.1.2
scipy==1.4.1
tqdm==4.46.1
